refacto : moved modules into a modules folder
