+++ api/src/db/data-source.ts
@@ -1,11 +1,11 @@
 import { DataSource } from "typeorm";
 import * as dotenv from "dotenv";
-import { Team } from "../team/team.entity";
-import { Competition } from "../competition/competition.entity";
-import { Jury } from "../jury/jury.entity";
-import { User } from "../user/user.entity";
-import { Role } from "../role/role.entity";
-import { Session } from "../session/session.entity";
+import { Team } from "../modules/team/team.entity";
+import { Competition } from "../modules/competition/competition.entity";
+import { Jury } from "../modules/jury/jury.entity";
+import { User } from "../modules/user/user.entity";
+import { Role } from "../modules/role/role.entity";
+import { Session } from "../modules/session/session.entity";
 
 dotenv.config();
 const { DB_FILE } = process.env;
