+++ api/src/modules/competition/competition.entity.ts
