+++ api/src/modules/competition/competition.resolver.ts
