+++ api/src/modules/jury/jury.entity.ts
