+++ api/src/modules/jury/jury.resolver.ts
