+++ api/src/modules/session/session.entity.ts
